@@ -33,11 +33,11 @@
     <description>Dubbo Task - Shop - Comment Gray</description>
 
     <properties>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
-        <dubbo.version>3.3.1</dubbo.version>
+        <dubbo.version>3.2.16</dubbo.version>
         <nacos.version>2.2.0</nacos.version>
     </properties>
 
@@ -64,6 +64,18 @@
         </dependency>
 
         <dependency>
+            <groupId>org.apache.dubbo</groupId>
+            <artifactId>dubbo-spring-boot-tracing-otel-otlp-starter</artifactId>
+            <version>${dubbo.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.dubbo</groupId>
+            <artifactId>dubbo-spring-boot-observability-starter</artifactId>
+            <version>${dubbo.version}</version>
+        </dependency>
+
+        <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter</artifactId>
             <exclusions>
@@ -86,6 +98,12 @@
         </dependency>
     </dependencies>
 
+    <repositories>
+        <repository>
+            <id>alibaba-maven</id>
+            <url>https://maven.aliyun.com/repository/public</url>
+        </repository>
+    </repositories>
     <build>
         <plugins>
             <plugin>
@@ -99,10 +117,10 @@
             <plugin>
                 <groupId>com.google.cloud.tools</groupId>
                 <artifactId>jib-maven-plugin</artifactId>
-                <version>3.2.1</version>
+                <version>3.4.5</version>
                 <configuration>
                     <to>
-                        <image>apache/dubbo-demo:dubbo-samples-shop-comment_gray_0.0.1</image>
+                        <image>apache/dubbo-demo:dubbo-samples-shop-comment_gray_0.0.2</image>
                     </to>
                     <container>
                         <jvmFlags>
