@@ -21,12 +21,12 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import org.apache.dubbo.config.spring.context.annotation.EnableDubbo;
 
-import org.junit.Test;
 import org.junit.jupiter.api.Assertions;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.web.client.RestClient;
 import org.springframework.web.client.RestClientResponseException;
 
@@ -35,7 +35,7 @@ import java.util.Base64;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @EnableDubbo
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 public class ResourceServerIT {
 
     //    @LocalServerPort
@@ -51,48 +51,48 @@ public class ResourceServerIT {
     private static final String HOST = System.getProperty("resource.address", "localhost");
 
     @Test
-        public void testGetUserEndpoint() {
-            String credentials = clientId + ":" + clientSecret;
-            String encodedCredentials = Base64.getEncoder().encodeToString(credentials.getBytes());
+    public void testGetUserEndpoint() {
+        String credentials = clientId + ":" + clientSecret;
+        String encodedCredentials = Base64.getEncoder().encodeToString(credentials.getBytes());
 
-            // build RestClient request
-            RestClient restClient = RestClient.builder().build();
-            String url = "http://"+ OAUTH2HOST + ":9000/oauth2/token";
+        // build RestClient request
+        RestClient restClient = RestClient.builder().build();
+        String url = "http://"+ OAUTH2HOST + ":9000/oauth2/token";
 
+        try {
+            // make a post request
+            String response = restClient.post()
+                    .uri(url)
+                    .header(HttpHeaders.AUTHORIZATION, "Basic " + encodedCredentials)
+                    .header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+                    .body("grant_type=client_credentials&scope=read")
+                    .retrieve()
+                    .body(String.class);
+
+            ObjectMapper objectMapper = new ObjectMapper();
+            JsonNode jsonNode = objectMapper.readTree(response);
+            String accessToken = jsonNode.get("access_token").asText();
+
+            // Use the access token to authenticate the request to the /user endpoint
+            assert accessToken != null;
+            String userUrl = "http://" + HOST + ":50051/hello/sayHello/World";
             try {
-                // make a post request
-                String response = restClient.post()
-                        .uri(url)
-                        .header(HttpHeaders.AUTHORIZATION, "Basic " + encodedCredentials)
-                        .header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_FORM_URLENCODED_VALUE)
-                        .body("grant_type=client_credentials&scope=read")
+                String userResponse = restClient.get()
+                        .uri(userUrl)
+                        .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
                         .retrieve()
                         .body(String.class);
 
-                ObjectMapper objectMapper = new ObjectMapper();
-                JsonNode jsonNode = objectMapper.readTree(response);
-                String accessToken = jsonNode.get("access_token").asText();
-
-                // Use the access token to authenticate the request to the /user endpoint
-                assert accessToken != null;
-                String userUrl = "http://" + HOST + ":50051/hello/sayHello/World";
-                try {
-                    String userResponse = restClient.get()
-                            .uri(userUrl)
-                            .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-                            .retrieve()
-                            .body(String.class);
-
-                    assertEquals("\"Hello, World\"", userResponse, "error");
-                } catch (RestClientResponseException e) {
-                    System.err.println("Error Response: " + e.getResponseBodyAsString());
-                    Assertions.fail("Request failed with response: " + e.getResponseBodyAsString());
-                }
-
-            } catch (JsonProcessingException e) {
-                throw new RuntimeException(e);
+                assertEquals("\"Hello, World\"", userResponse, "error");
+            } catch (RestClientResponseException e) {
+                System.err.println("Error Response: " + e.getResponseBodyAsString());
+                Assertions.fail("Request failed with response: " + e.getResponseBodyAsString());
             }
+
+        } catch (JsonProcessingException e) {
+            throw new RuntimeException(e);
         }
+    }
 
     @Test
     public void testGetUserEndpointWithInvalidToken() {
