@@ -42,10 +42,6 @@ import java.util.stream.Collectors;
  * <p>
  * NOTE: at least an external standalone server (if not an ensemble) are recommended, even for
  * {@link org.springframework.xd.dirt.server.singlenode.SingleNodeApplication}
- *
- * @author Patrick Peralta
- * @author Mark Fisher
- * @author David Turanski
  */
 public class EmbeddedZooKeeper implements SmartLifecycle {
 
@@ -243,7 +239,7 @@ public class EmbeddedZooKeeper implements SmartLifecycle {
                 zkServer = new ZooKeeperServerMain();
                 ServerConfig configuration = new ServerConfig();
                 configuration.readFrom(quorumPeerConfig);
-                
+
                 System.setProperty("zookeeper.admin.enableServer", "false");
 
                 zkServer.runFromConfig(configuration);
