@@ -18,22 +18,59 @@
 package org.apache.dubbo.rest.demo;
 
 import org.apache.dubbo.config.annotation.DubboService;
+import org.apache.dubbo.remoting.http12.HttpResult;
+
+import java.util.Collections;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.IntStream;
 
 @DubboService
 public class DemoServiceImpl implements DemoService {
 
+    private static final StringBuffer finalString = new StringBuffer();
+
+    private static final Map<String, List<String>> finalMap = new LinkedHashMap<>();
+
+    static {
+        IntStream.range(0, 20000).forEach(i -> finalString.append(i).append("Hello"));
+        finalMap.put("large-size-response-header", Collections.singletonList(finalString.toString()));
+    }
+
     @Override
     public String hello(String name) {
-        return "Hello " + name;
+        return finalString + "Hello " + name;
     }
 
     @Override
     public String hello(User user, int count) {
-        return "Hello " + user.getTitle() + ". " + user.getName() + ", " + count;
+        return finalString + "Hello " + user.getTitle() + ". " + user.getName() + ", " + count;
     }
 
     @Override
     public String helloUser(User user) {
-        return "Hello " + user.getTitle() + ". " + user.getName();
+        return finalString + "Hello " + user.getTitle() + ". " + user.getName();
+    }
+
+    @Override
+    public HttpResult<String> helloH2cWithExceededLargeSizeHeaderResp() {
+        return new HttpResult<>() {
+
+            @Override
+            public int getStatus() {
+                return 200;
+            }
+
+            @Override
+            public Map<String, List<String>> getHeaders() {
+                return finalMap;
+            }
+
+            @Override
+            public String getBody() {
+                return "Client should not get this because response header size exceeds limitation!";
+            }
+        };
     }
 }
